refactor(SystemStatus): consolidate code
