+++ lib/dotcom/system_status/subway.ex
@@ -8,7 +8,7 @@ defmodule Dotcom.SystemStatus.Subway do
 
   @type status_time() :: :current | {:future, DateTime.t()}
 
-  @type status_t() :: :normal | Dotcom.SystemStatus.Alerts.service_effect_t()
+  @type status_t() :: :normal | Dotcom.Alerts.service_effect_t()
 
   @type status_entry() :: %{
           alerts: [Alert.t()],
